directory/BeeSafe_WebApp/requirements.txt
altair==4.1.0
appnope==0.1.0
astor==0.8.1
astropy==4.0.1.post1
attrs==19.3.0
awscli==1.18.72
backcall==0.1.0
base58==2.0.1
bleach==3.1.4
blinker==1.4
blis==0.4.1
boto==2.49.0
boto3==1.13.22
botocore==1.16.22
cachetools==4.1.0
catalogue==1.0.0
catboost==0.20.2
certifi==2020.4.5.2
chardet==3.0.4
chart-studio==1.1.0
click==7.1.2
colorama==0.4.3
colorlover==0.3.0
combo==0.1.0
confuse==1.1.0
cufflinks==0.17.0
cycler==0.10.0
cymem==2.0.3
datefinder==0.7.0
DateTime==4.3
deap==1.3.1
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
en-core-web-sm==2.2.5
entrypoints==0.3
enum-compat==0.0.3
Flask==1.1.2
funcy==1.14
future==0.18.2
gensim==3.8.3
graphviz==0.14
htmlmin==0.1.12
idna==2.9
imbalanced-learn==0.6.2
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
kmodes==0.10.1
lightgbm==2.3.1
llvmlite==0.32.1
MarkupSafe==1.1.1
matplotlib==3.1.3
missingno==0.4.2
mistune==0.8.4
mkl-fft==1.0.15
mkl-service==2.3.0
mlxtend==0.17.2
more-itertools==8.3.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.6
nltk==3.5
notebook==6.0.3
numba==0.49.1
numexpr==2.7.1
numpy==1.18.1
olefile==0.46
packaging==20.4
pandas==1.0.3
pandas-profiling==2.3.0
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
phik==0.10.0
pickleshare==0.7.5
Pillow==7.1.2
plac==1.1.3
plotly==4.8.1
pluggy==0.13.1
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.12.2
psycopg2==2.8.4
ptyprocess==0.6.0
py==1.8.1
py4j==0.10.7
pyasn1==0.4.8
pycaret==1.0.0
pydeck==0.4.0b2
Pygments==2.6.1
pyLDAvis==2.1.2
pyod==0.8.0
pyparsing==2.4.7
pyrsistent==0.16.0
pyspark==2.4.4
pytest==5.4.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==4.7.3
QtPy==1.9.0
regex==2020.5.14
requests==2.23.0
retrying==1.3.3
rsa==3.4.2
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
shap==0.32.1
six==1.14.0
smart-open==2.0.0
smogn==0.1.1
spacy==2.2.4
SQLAlchemy==1.3.16
SQLAlchemy-Utils==0.36.5
srsly==1.0.2
stopit==1.1.2
streamlit==0.61.0
suod==0.0.4
tbb==2019.0
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
thinc==7.4.0
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.10.0
tornado==5.1.1
TPOT==0.11.5
tqdm==4.46.1
traitlets==4.3.3
tzlocal==2.1
umap-learn==0.4.3
update-checker==0.17
urllib3==1.25.9
validators==0.15.0
vega-datasets==0.7.0
wasabi==0.6.0
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wordcloud==1.7.0
xgboost==0.90
yellowbrick==1.0.1
zipp==3.1.0
zope.interface==5.1.0
